Refactor for addressing issues in boundary_face and subgrid model.
This is a significant refactor changing the stencil structure of
overland conductivity models that need boundary faces.  Particularly,
some of the quantities in that term only make sense on cells, not on
faces.  Previously, we had hacked those terms to exist on faces, but
it made things harder in other ways (e.g. #106 ).  Now, we have a
mixed model where some of those quantities are defined on
faces (e.g. primary variables such as p or delta) but others are
not (e.g. manning coefficient, slope magnitudes, etc).  The natural
thing is to take the interior cell's value on boundary faces in this
case.

Also cleans up documentation of the subgrid model.  New tests are in
progress/added on branches and will be included in this PR.