Cleanup, renaming some MPCs for transport, and fixes bug in mpc_surface
This does some code cleanup and updates some MPC names to be a bit
more typical of ATS naming conventions.

It also fixes a bug in mpc_surface where the derivative was required,
but not used (which caused a seg fault because no metadata was
provided for that field).

This gets 06_energy/freeze_thaw_surface to pass.