changes the algorithm to set a max pc at the plant collar
This removes the use of a linear beta function for max pc e.g. CLM,
and is no longer needed with the solve for capillary pressure.

The result is a change in the parameter name in trf relperm, and
rescales units to match relperm and permeability in the correct units
given ATS relative perm conventions.