Changes design of viewComponent
CompositeVector&::viewComponent previously accepted an
execution or memory space as a template argument.  This is bad -- a CV
is a collection of Tpetra::MultiVectors under the hood, which are
really dual views.  So not all execution/memory spaces are valid --
only those consistent with the dual view's t_dev or t_host.

This led to lots of problems, particularly where

viewComponent<HostSpace>()

is used -- this is not necessarily a valid space for e.g. CUDA, where
it needs to be SpecialHostSpace or MirrorSpace.

Instead, these are now all templated with MemSpace_kind for HOST or
DEVICE views.