fixes transport for parallel runs (#306)
Careful pass through of parallelization logic of transport, reactive transport, creating contexts for each component of the Advance.

Note bugs here were not caught because of poor management of Epetra MultiVector return codes.  In the future, we should be more careful to catch nonzero return codes, as they will otherwise silently do nothing if, for instance, one vector is ghosted and the other is not.